requirements.txt
duckduckgo_search==7.2.1
groq==0.15.0
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.33
langchain-google-genai==2.0.7
langchain-groq==0.2.3
langchain-ollama==0.2.2
langchain-openai==0.3.3
langchain-postgres==0.0.12
langchain-text-splitters==0.3.4
langgraph==0.2.62
langgraph-checkpoint==2.0.9
langgraph-sdk==0.1.51
langsmith==0.2.4
openai==1.61.0
pgvector==0.2.5
psycopg==3.2.4
psycopg-binary==3.2.4
psycopg-pool==3.2.4
pydantic==2.10.4
pypdf==5.2.0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
streamlit==1.41.1
wikipedia==1.4.0
yfinance==0.2.54
youtube-search==2.1.2
